painel_ponto_eletronico/requirements.txt
django
psycopg[binary]
paho-mqtt
python-dotenv